소수점 계산 수정

+++ app/src/main/java/com/example/calculator/AddOperation.kt
@@ -1,9 +1,9 @@
 package com.example.calculator
 
-class AddOperation(private val num1: Int, private val num2: Int) {
+class AddOperation(private val num1: Double, private val num2: Double) {
     private var operationResult = num1 + num2
 
-    fun operationReturn(): Int {
+    fun operationReturn(): Double {
         println("결과 : $operationResult")
         return operationResult
     }
